refactor: extract Inspector.tsx sub-components into components/inspector/ (#395)
- Extract shared types/helpers into inspector/types.ts
- Extract InfoTooltip into inspector/InfoTooltip.tsx
- Extract AIPanel into inspector/AIPanel.tsx
- Extract CorrectionsPanel into inspector/CorrectionsPanel.tsx (~376 lines)
- Extract StatsPanel into inspector/StatsPanel.tsx (~495 lines)
- Add barrel export via inspector/index.ts
- Inspector.tsx reduced from 1,441 to 406 lines (orchestrator only)
