+++ components/Inspector.tsx
@@ -1,97 +1,20 @@
 "use client";
 
-import { useState, useRef, useEffect, useMemo, ReactNode } from "react";
-import { Bot, AlertCircle, BarChart3, Edit2, X, History, RefreshCw, Settings, Info, ListFilter, XCircle, AlertTriangle, EyeOff } from "lucide-react";
+import { useState, useRef, useEffect } from "react";
+import { Bot, AlertCircle, BarChart3, Edit2, X, History } from "lucide-react";
 import clsx from "clsx";
 
 import { useEditorMode } from "@/contexts/EditorModeContext";
 import { localPreferences } from "@/lib/local-preferences";
-import { LINT_PRESETS, LINT_RULES_META } from "@/lib/linting/lint-presets";
-import { DEFAULT_POS_COLORS } from "@/packages/milkdown-plugin-japanese-novel/pos-highlight/pos-colors";
 import HistoryPanel from "./HistoryPanel";
+import AIPanel from "./inspector/AIPanel";
+import CorrectionsPanel from "./inspector/CorrectionsPanel";
+import StatsPanel from "./inspector/StatsPanel";
 
 import type { ProjectMode } from "@/lib/project-types";
-import type { LintIssue, Severity } from "@/lib/linting";
-
-type Tab = "ai" | "corrections" | "stats" | "history";
-
-const isValidTab = (value: string | null): value is Tab =>
-  value === "ai" || value === "corrections" || value === "stats" || value === "history";
-
-const MDI_EXTENSION = ".mdi";
-
-function getMdiExtension(name: string) {
-  if (name.toLowerCase().endsWith(MDI_EXTENSION)) {
-    return name.slice(name.length - MDI_EXTENSION.length);
-  }
-  return "";
-}
-
-function getBaseName(name: string) {
-  const extension = getMdiExtension(name);
-  return extension ? name.slice(0, -extension.length) : name;
-}
-
-interface InspectorProps {
-  className?: string;
-  compactMode?: boolean;
-  charCount?: number;
-  selectedCharCount?: number;
-  paragraphCount?: number;
-  fileName?: string;
-  isDirty?: boolean;
-  isSaving?: boolean;
-  lastSavedTime?: number | null;
-  onSaveFile?: () => void;
-  onFileNameChange?: (newName: string) => void;
-  sentenceCount?: number;
-  charTypeAnalysis?: {
-    kanji: number;
-    hiragana: number;
-    katakana: number;
-    other: number;
-    total: number;
-  };
-  charUsageRates?: {
-    kanjiRate: number;
-    hiraganaRate: number;
-    katakanaRate: number;
-  };
-  readabilityAnalysis?: {
-    score: number;
-    level: string;
-    avgSentenceLength: number;
-    avgPunctuationSpacing: number;
-  };
-  // 品詞着色設定
-  posHighlightEnabled?: boolean;
-  onPosHighlightEnabledChange?: (enabled: boolean) => void;
-  posHighlightColors?: Record<string, string>;
-  onOpenPosHighlightSettings?: () => void;
-  // 履歴復元コールバック（プロジェクトモード時に使用）
-  onHistoryRestore?: (content: string) => void;
-  // 現在開いているファイル名（履歴パネルの切り替え用）
-  activeFileName?: string;
-  // 現在のエディタ内容（履歴差分表示用）
-  currentContent?: string;
-  // エディタ領域で差分を表示するコールバック
-  onCompareInEditor?: (data: { snapshotContent: string; currentContent: string; label: string }) => void;
-  // リンティング結果
-  lintIssues?: LintIssue[];
-  onNavigateToIssue?: (issue: LintIssue) => void;
-  onApplyFix?: (issue: LintIssue) => void;
-  onIgnoreCorrection?: (issue: LintIssue, ignoreAll: boolean) => void;
-  onRefreshLinting?: () => void;
-  isLinting?: boolean;
-  // Active lint issue (cursor sync)
-  activeLintIssueIndex?: number | null;
-  // Settings integration
-  onOpenLintingSettings?: () => void;
-  onApplyLintPreset?: (presetId: string) => void;
-  activeLintPresetId?: string;
-  /** Monotonically increasing trigger to switch to the corrections tab from outside */
-  switchToCorrectionsTrigger?: number;
-}
+import type { InspectorProps } from "./inspector/types";
+import type { Tab } from "./inspector/types";
+import { isValidTab, getMdiExtension, getBaseName } from "./inspector/types";
 
 export default function Inspector({
   className,
@@ -190,6 +113,7 @@ export default function Inspector({
       setActiveTab("ai");
     }
   }, [isProject, activeTab]);
+
   const [isEditingFileName, setIsEditingFileName] = useState(false);
   const [editedBaseName, setEditedBaseName] = useState(() => getBaseName(fileName));
   const inputRef = useRef<HTMLInputElement>(null);
@@ -480,962 +404,3 @@ export default function Inspector({
     </aside>
   );
 }
-
-function AIPanel() {
-  return (
-    <div className="space-y-4">
-      <div className="bg-accent-light rounded-lg p-4 border border-border">
-        <div className="flex items-start gap-3">
-          <Bot className="w-5 h-5 text-accent mt-0.5" />
-          <div>
-            <h3 className="text-sm font-medium text-foreground mb-1">AI アシスタント</h3>
-            <p className="text-xs text-foreground-tertiary">
-              この機能は現在開発中です。今後のアップデートをお待ちください。
-            </p>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-/** Severity filter options for the corrections panel */
-type SeverityFilter = "all" | Severity;
-
-/** Extended issue with original text from the document */
-interface EnrichedLintIssue extends LintIssue {
-  originalText?: string;
-}
-
-interface CorrectionsPanelProps {
-  posHighlightEnabled: boolean;
-  onPosHighlightEnabledChange?: (enabled: boolean) => void;
-  posHighlightColors?: Record<string, string>;
-  onOpenPosHighlightSettings?: () => void;
-  lintIssues: (LintIssue | EnrichedLintIssue)[];
-  onNavigateToIssue?: (issue: LintIssue) => void;
-  onApplyFix?: (issue: LintIssue) => void;
-  onIgnoreCorrection?: (issue: LintIssue, ignoreAll: boolean) => void;
-  onRefreshLinting?: () => void;
-  isLinting?: boolean;
-  activeLintIssueIndex?: number | null;
-  onOpenLintingSettings?: () => void;
-  onApplyLintPreset?: (presetId: string) => void;
-  activeLintPresetId?: string;
-}
-
-/** Returns the display color class for a severity level */
-function severityColor(severity: Severity): string {
-  switch (severity) {
-    case "error":
-      return "bg-error";
-    case "warning":
-      return "bg-warning";
-    case "info":
-      return "bg-info";
-  }
-}
-
-/** Get rule display name from metadata */
-function getRuleName(ruleId: string): string {
-  const meta = LINT_RULES_META.find(r => r.id === ruleId);
-  return meta?.nameJa ?? ruleId;
-}
-
-/** POS legend items (main categories only, compact) */
-const POS_LEGEND_ITEMS = [
-  { key: "名詞", label: "名詞" },
-  { key: "動詞", label: "動詞" },
-  { key: "形容詞", label: "形容詞" },
-  { key: "副詞", label: "副詞" },
-  { key: "助詞", label: "助詞" },
-  { key: "助動詞", label: "助動詞" },
-  { key: "接続詞", label: "接続詞" },
-  { key: "連体詞", label: "連体詞" },
-  { key: "感動詞", label: "感動詞" },
-  { key: "記号", label: "記号" },
-];
-
-function CorrectionsPanel({
-  posHighlightEnabled,
-  onPosHighlightEnabledChange,
-  posHighlightColors = {},
-  onOpenPosHighlightSettings,
-  lintIssues,
-  onNavigateToIssue,
-  onApplyFix,
-  onIgnoreCorrection,
-  onRefreshLinting,
-  isLinting = false,
-  activeLintIssueIndex,
-  onOpenLintingSettings,
-  onApplyLintPreset,
-  activeLintPresetId = "",
-}: CorrectionsPanelProps) {
-  const [severityFilter, setSeverityFilter] = useState<SeverityFilter>("all");
-  const issueListRef = useRef<HTMLDivElement>(null);
-
-  const filteredIssues = useMemo(() =>
-    severityFilter === "all"
-      ? lintIssues
-      : lintIssues.filter((issue) => issue.severity === severityFilter),
-    [lintIssues, severityFilter]
-  );
-
-  // Find active index within filtered list
-  const activeFilteredIndex = useMemo(() => {
-    if (activeLintIssueIndex == null) return null;
-    const activeIssue = lintIssues[activeLintIssueIndex];
-    if (!activeIssue) return null;
-    return filteredIssues.indexOf(activeIssue);
-  }, [activeLintIssueIndex, lintIssues, filteredIssues]);
-
-  // Auto-scroll active card into view
-  useEffect(() => {
-    if (activeFilteredIndex == null || activeFilteredIndex < 0) return;
-    const container = issueListRef.current;
-    if (!container) return;
-    const card = container.querySelector(`[data-issue-index="${activeFilteredIndex}"]`);
-    if (card) {
-      card.scrollIntoView({ behavior: "smooth", block: "nearest" });
-    }
-  }, [activeFilteredIndex]);
-
-  const filterOptions: { value: SeverityFilter; label: string; icon: ReactNode }[] = [
-    { value: "all", label: "全て", icon: <ListFilter className="w-3.5 h-3.5" /> },
-    { value: "error", label: "エラー", icon: <XCircle className="w-3.5 h-3.5" /> },
-    { value: "warning", label: "警告", icon: <AlertTriangle className="w-3.5 h-3.5" /> },
-    { value: "info", label: "情報", icon: <Info className="w-3.5 h-3.5" /> },
-  ];
-
-  return (
-    <div className="space-y-3">
-      {/* POS highlight toggle (existing feature) */}
-      <div className="bg-background-secondary rounded-lg p-3 border border-border">
-        <div className="flex items-center justify-between">
-          <div>
-            <h4 className="text-sm font-medium text-foreground">構文ハイライト</h4>
-            <p className="text-xs text-foreground-tertiary mt-0.5">
-              動詞・助詞などを色分け表示
-            </p>
-          </div>
-          <button
-            onClick={() => onPosHighlightEnabledChange?.(!posHighlightEnabled)}
-            className={clsx(
-              "relative inline-flex h-6 w-11 items-center rounded-full transition-colors",
-              posHighlightEnabled ? "bg-accent" : "bg-foreground-muted"
-            )}
-          >
-            <span
-              className={clsx(
-                "inline-block h-4 w-4 transform rounded-full bg-white transition-transform shadow-sm",
-                posHighlightEnabled ? "translate-x-6" : "translate-x-1"
-              )}
-            />
-          </button>
-        </div>
-        {posHighlightEnabled && (
-          <div className="mt-2 space-y-2">
-            <div className="flex flex-wrap gap-x-3 gap-y-1">
-              {POS_LEGEND_ITEMS.map(({ key, label }) => (
-                <span key={key} className="inline-flex items-center gap-1 text-xs text-foreground-secondary">
-                  <span
-                    className="inline-block w-2.5 h-2.5 rounded-sm"
-                    style={{ backgroundColor: posHighlightColors[key] || DEFAULT_POS_COLORS[key] || "#000" }}
-                  />
-                  {label}
-                </span>
-              ))}
-            </div>
-            {onOpenPosHighlightSettings && (
-              <button
-                onClick={onOpenPosHighlightSettings}
-                className="text-xs text-accent hover:text-accent-hover hover:underline transition-colors"
-              >
-                色の設定を変更
-              </button>
-            )}
-          </div>
-        )}
-      </div>
-
-      {/* Header: issue count + controls */}
-      <div className="space-y-1.5">
-        <div className="flex items-center justify-between">
-          <div className="flex items-center gap-1">
-            <h3 className="text-sm font-medium text-foreground-secondary">検出結果</h3>
-            <span className="text-xs text-foreground-tertiary">
-              {lintIssues.length}件
-            </span>
-          </div>
-          <div className="flex items-center gap-1">
-            {/* Refresh button */}
-            {onRefreshLinting && (
-              <button
-                onClick={onRefreshLinting}
-                disabled={isLinting}
-                className={clsx(
-                  "p-1 rounded transition-colors",
-                  isLinting
-                    ? "text-accent cursor-wait"
-                    : "text-foreground-tertiary hover:text-foreground-secondary hover:bg-hover"
-                )}
-                title="全文を再検査"
-                aria-label="全文を再検査"
-              >
-                <RefreshCw className={clsx("w-3.5 h-3.5", isLinting && "animate-spin")} />
-              </button>
-            )}
-            {onOpenLintingSettings && (
-              <button
-                onClick={onOpenLintingSettings}
-                className="p-1 text-foreground-tertiary hover:text-foreground-secondary hover:bg-hover rounded transition-colors"
-                title="校正設定を開く"
-              >
-                <Settings className="w-3.5 h-3.5" />
-              </button>
-            )}
-          </div>
-          {onApplyLintPreset && (
-            <select
-              value={activeLintPresetId}
-              onChange={(e) => {
-                if (e.target.value) {
-                  onApplyLintPreset(e.target.value);
-                }
-              }}
-              className="text-xs px-1.5 py-0.5 border border-border-secondary rounded bg-background text-foreground focus:outline-none focus:ring-1 focus:ring-accent"
-              title="プリセットを適用"
-            >
-              {!activeLintPresetId && <option value="">カスタム</option>}
-              {Object.entries(LINT_PRESETS).map(([id, preset]) => (
-                <option key={id} value={id}>{preset.nameJa}</option>
-              ))}
-            </select>
-          )}
-        </div>
-      </div>
-
-      {/* Severity filter buttons */}
-      <div className="flex gap-1">
-        {filterOptions.map((option) => (
-          <button
-            key={option.value}
-            onClick={() => setSeverityFilter(option.value)}
-            title={option.label}
-            aria-label={option.label}
-            aria-pressed={severityFilter === option.value}
-            className={clsx(
-              "flex-1 flex items-center justify-center px-2 py-1.5 rounded transition-colors",
-              severityFilter === option.value
-                ? "bg-accent text-white"
-                : "bg-background-secondary text-foreground-tertiary hover:text-foreground-secondary border border-border"
-            )}
-          >
-            {option.icon}
-          </button>
-        ))}
-      </div>
-
-      {/* Issue list */}
-      {filteredIssues.length === 0 ? (
-        <div className="pt-4 text-center">
-          <p className="text-sm text-foreground-tertiary">問題は検出されませんでした</p>
-        </div>
-      ) : (
-        <div ref={issueListRef} className="space-y-1.5">
-          {filteredIssues.map((issue, index) => {
-            const enriched = issue as EnrichedLintIssue;
-            const isActive = activeFilteredIndex === index;
-            const hasOriginal = !!enriched.originalText;
-            const hasFix = !!issue.fix;
-
-            return (
-              <div
-                key={`${issue.ruleId}-${issue.from}-${issue.to}-${index}`}
-                data-issue-index={index}
-                className={clsx(
-                  "rounded-lg border transition-colors",
-                  isActive
-                    ? "border-accent bg-accent/5"
-                    : "border-border bg-background-secondary hover:border-border-secondary"
-                )}
-              >
-                {/* Actions: top-right corner */}
-                <div className="flex items-center gap-1 float-right ml-2 mt-2 mr-2">
-                  <InfoTooltip
-                    content={`${issue.messageJa}${issue.reference ? `\n${issue.reference.standard}${issue.reference.section ? ` ${issue.reference.section}` : ""}` : ""}\n[${getRuleName(issue.ruleId)}]`}
-                    className="text-foreground-tertiary hover:text-foreground-secondary"
-                  >
-                    <Info className="w-3.5 h-3.5" />
-                  </InfoTooltip>
-                  {hasFix && (
-                    <span
-                      role="button"
-                      tabIndex={0}
-                      onClick={(e) => {
-                        e.stopPropagation();
-                        onApplyFix?.(issue);
-                      }}
-                      onKeyDown={(e) => {
-                        if (e.key === "Enter" || e.key === " ") {
-                          e.stopPropagation();
-                          onApplyFix?.(issue);
-                        }
-                      }}
-                      className="inline-flex items-center px-1.5 py-0.5 text-xs font-medium text-accent hover:text-accent-hover bg-accent/10 hover:bg-accent/20 rounded transition-colors cursor-pointer"
-                    >
-                      置換
-                    </span>
-                  )}
-                  {onIgnoreCorrection && (
-                    <InfoTooltip
-                      content="この指摘を無視（右クリックで同じ指摘をすべて無視）"
-                      className="text-foreground-tertiary hover:text-foreground-secondary"
-                    >
-                      <span
-                        role="button"
-                        tabIndex={0}
-                        aria-label="この指摘を無視"
-                        onClick={(e) => {
-                          e.stopPropagation();
-                          onIgnoreCorrection(issue, false);
-                        }}
-                        onContextMenu={(e) => {
-                          e.preventDefault();
-                          e.stopPropagation();
-                          onIgnoreCorrection(issue, true);
-                        }}
-                        onKeyDown={(e) => {
-                          if (e.key === "Enter" || e.key === " ") {
-                            e.stopPropagation();
-                            onIgnoreCorrection(issue, false);
-                          }
-                        }}
-                        className="inline-flex items-center p-0.5 rounded transition-colors cursor-pointer hover:bg-hover"
-                      >
-                        <EyeOff className="w-3.5 h-3.5" />
-                      </span>
-                    </InfoTooltip>
-                  )}
-                </div>
-                {/* Main row: click to navigate */}
-                <button
-                  type="button"
-                  onClick={() => onNavigateToIssue?.(issue)}
-                  className="w-full text-left px-3 py-2"
-                >
-                  <div className="flex items-start gap-2">
-                    {/* Severity dot */}
-                    <span
-                      className={clsx(
-                        "w-2 h-2 rounded-full shrink-0 mt-1.5",
-                        severityColor(issue.severity)
-                      )}
-                      title={issue.severity}
-                    />
-                    {/* Content: original → replacement or just message */}
-                    <div className="flex-1 min-w-0">
-                      {hasOriginal && hasFix ? (
-                        <p className="text-sm text-foreground leading-snug">
-                          <span className="text-foreground-tertiary line-through">{enriched.originalText}</span>
-                          <span className="text-foreground-tertiary mx-1">→</span>
-                          <span className="text-foreground font-medium">{issue.fix!.replacement}</span>
-                        </p>
-                      ) : hasOriginal ? (
-                        <p className="text-sm text-foreground leading-snug">
-                          {enriched.originalText}
-                        </p>
-                      ) : (
-                        <p className="text-sm text-foreground leading-snug">
-                          {issue.messageJa}
-                        </p>
-                      )}
-                      {/* Secondary line: show message when we have original text but no fix */}
-                      {hasOriginal && !hasFix && (
-                        <p className="text-xs text-foreground-tertiary mt-0.5">
-                          {issue.messageJa}
-                        </p>
-                      )}
-                    </div>
-                  </div>
-                </button>
-              </div>
-            );
-          })}
-        </div>
-      )}
-    </div>
-  );
-}
-
-
-// 情報アイコン用のツールチップ
-function InfoTooltip({ content, className, children }: { content: string; className?: string; children: ReactNode }) {
-  const [isVisible, setIsVisible] = useState(false);
-  const [tooltipPos, setTooltipPos] = useState({ top: 0, left: 0, placement: 'top' as 'top' | 'bottom' });
-  const ref = useRef<HTMLSpanElement>(null);
-  const tooltipRef = useRef<HTMLDivElement>(null);
-  
-  const handleMouseEnter = () => {
-    if (ref.current) {
-      const rect = ref.current.getBoundingClientRect();
-      setTooltipPos({
-        top: rect.top - 8,
-        left: rect.left + rect.width / 2,
-        placement: 'top',
-      });
-    }
-    setIsVisible(true);
-  };
-
-  useEffect(() => {
-    if (!isVisible || !tooltipRef.current || !ref.current) return;
-    const tooltipRect = tooltipRef.current.getBoundingClientRect();
-    const iconRect = ref.current.getBoundingClientRect();
-    const margin = 8;
-
-    let left = iconRect.left + iconRect.width / 2;
-    const minLeft = margin + tooltipRect.width / 2;
-    const maxLeft = window.innerWidth - margin - tooltipRect.width / 2;
-    if (left < minLeft) left = minLeft;
-    if (left > maxLeft) left = maxLeft;
-
-    let top = iconRect.top - 8;
-    let placement: 'top' | 'bottom' = 'top';
-    if (top - tooltipRect.height < margin) {
-      top = iconRect.bottom + 8;
-      placement = 'bottom';
-    }
-
-    setTooltipPos({ top, left, placement });
-  }, [isVisible]);
-  
-  return (
-    <span 
-      ref={ref}
-      className={clsx("info-tooltip-wrapper cursor-help", className)}
-      onMouseEnter={handleMouseEnter}
-      onMouseLeave={() => setIsVisible(false)}
-    >
-      {children}
-      {isVisible && (
-        <span 
-          ref={tooltipRef}
-          className="info-tooltip-content"
-          style={{
-            position: 'fixed',
-            top: `${tooltipPos.top}px`,
-            left: `${tooltipPos.left}px`,
-            transform: tooltipPos.placement === 'top'
-              ? 'translateX(-50%) translateY(-100%)'
-              : 'translateX(-50%) translateY(0)'
-          }}
-        >
-          {content}
-        </span>
-      )}
-    </span>
-  );
-}
-
-function StatsPanel({
-  charCount,
-  selectedCharCount,
-  paragraphCount,
-  manuscriptPages,
-  sentenceCount = 0,
-  charTypeAnalysis,
-  charUsageRates,
-  readabilityAnalysis,
-}: {
-  charCount: number;
-  selectedCharCount: number;
-  paragraphCount: number;
-  manuscriptPages: number;
-  sentenceCount?: number;
-  charTypeAnalysis?: {
-    kanji: number;
-    hiragana: number;
-    katakana: number;
-    other: number;
-    total: number;
-  };
-  charUsageRates?: {
-    kanjiRate: number;
-    hiraganaRate: number;
-    katakanaRate: number;
-  };
-  readabilityAnalysis?: {
-    score: number;
-    level: string;
-    avgSentenceLength: number;
-    avgPunctuationSpacing: number;
-  };
-}) {
-  // 選択範囲の分析かどうか
-  const isSelection = selectedCharCount > 0;
-  const activeCharCount = isSelection ? selectedCharCount : charCount;
-
-  // 1. 文字数の詳細（概算）
-  // 本文文字数を概算（空白・約物などを除外）
-  // 注意: ここでは簡易推定。厳密にはエディタの生テキストを用いる
-  // 約物は全体の 10-15% 程度と仮定
-  const estimatedPunctuation = Math.floor(activeCharCount * 0.12);
-  const pureTextCount = activeCharCount - estimatedPunctuation;
-  const punctuationRatio = activeCharCount > 0 ? (estimatedPunctuation / activeCharCount * 100).toFixed(1) : '0.0';
-
-   // 文体の判定
-   let styleHint = '';
-   const ratio = parseFloat(punctuationRatio);
-   if (ratio > 15) {
-     styleHint = '会話文が中心';
-   } else if (ratio < 8) {
-     styleHint = '地の文が中心';
-   } else {
-     styleHint = 'バランス型';
-   }
-
-  // 2. 段落構成
-  const avgParagraphLength = paragraphCount > 0 ? Math.floor(activeCharCount / paragraphCount) : 0;
-
-  let paragraphWarning = '';
-  if (avgParagraphLength >= 300) {
-    paragraphWarning =
-        '一段落に含まれる情報量がやや多いようです。内容のまとまりごとに区切ると、読みやすさが向上するかもしれません。';
-
-  } else if (avgParagraphLength >= 200) {
-    paragraphWarning =
-        '読み応えのある段落構成です。公的文書や解説文としては安定していますが、スマホでは少し長く感じられる場合があります。';
-
-  } else if (avgParagraphLength >= 120) {
-    paragraphWarning =
-        '段落の長さは標準的で、エッセイや一般的な文章に適した構成です。落ち着いて読み進められます。';
-
-  } else if (avgParagraphLength >= 80) {
-    paragraphWarning =
-        '小説向きの自然な段落長です。文章のリズムと情報量のバランスが保たれています。';
-
-  } else if (avgParagraphLength > 0) {
-    paragraphWarning =
-        '段落がコンパクトで、テンポよく読めます。会話やスマホでの読書に向いた構成です。';
-  }
-
-  // 3. 読了時間（シーン別の目安）
-  const calculateReadTime = (charsPerMinute: number) => {
-    if (activeCharCount === 0) return '0秒';
-    const minutes = Math.floor(activeCharCount / charsPerMinute);
-    const seconds = Math.round((activeCharCount % charsPerMinute) / charsPerMinute * 60);
-
-    if (minutes === 0) {
-      return `${seconds}秒`;
-    } else if (seconds === 0) {
-      return `${minutes}分`;
-    } else {
-      return `${minutes}分${seconds}秒`;
-    }
-  };
-
-   const fastReadTime = calculateReadTime(900);    // 速読
-   const normalReadTime = calculateReadTime(500);  // 通常読書
-   const deepReadTime = calculateReadTime(250);    // 精読
-
-    const getReadabilityLevelLabel = (level?: string) => {
-      switch (level) {
-        case 'easy':
-          return 'やさしい';
-        case 'normal':
-          return '標準';
-        case 'difficult':
-          return '難しい';
-        default:
-          return '未分析';
-      }
-    };
-
-    return (
-      <div className="space-y-3 stats-panel">
-        {/* 0. 原稿用紙枚数（全体のみ表示、トップに配置） */}
-        {!isSelection && (
-          <div className="bg-background-secondary rounded-lg p-3 border border-border flex items-center justify-between">
-            <div>
-              <p className="text-xs text-foreground-tertiary font-medium mb-1 flex items-center">
-                <InfoTooltip content="400字詰め原稿用紙に換算した枚数">
-                  原稿用紙
-                </InfoTooltip>
-              </p>
-              <p className="text-xs text-foreground-tertiary">400字詰原稿用紙</p>
-            </div>
-            <span className="text-sm font-bold text-foreground">{manuscriptPages}枚</span>
-          </div>
-        )}
-
-        {/* 見出し: 分析対象を動的に表示 */}
-        <div className="flex items-center justify-between">
-          <h3 className="stats-header">
-            {isSelection ? '選択範囲の分析' : '全体の統計'}
-          </h3>
-          {isSelection && (
-            <span className="text-xs px-2 py-1 rounded-full bg-accent/20 text-accent font-medium">
-              選択中
-            </span>
-          )}
-        </div>
-
-        {/* 1. 可読性分析 (Readability) - Featured Section */}
-        {readabilityAnalysis && !isSelection && (
-          <div className="bg-background-secondary rounded-lg p-4 border border-border">
-            <h4 className="text-xs font-medium text-foreground-tertiary uppercase tracking-wide mb-3">
-               読みやすさ
-            </h4>
-            <div className="space-y-2">
-              <div className="flex items-center justify-between gap-2">
-                <div className="flex items-center gap-1 min-w-0">
-                  <InfoTooltip
-                    content="文章の読みやすさを100点満点で評価。文の長さや句読点の配置から算出"
-                    className="text-sm text-foreground-secondary whitespace-nowrap"
-                  >
-                    スコア
-                  </InfoTooltip>
-                </div>
-                <div className="flex items-baseline gap-1 flex-shrink-0">
-                  <span className="text-xl font-bold text-foreground">{readabilityAnalysis.score}</span>
-                  <span className="text-xs text-foreground-tertiary">/100</span>
-                </div>
-              </div>
-               <div className="w-full h-2 bg-background rounded-full overflow-hidden border border-border-secondary">
-                 <div
-                   className="h-full transition-all"
-                   style={{
-                     width: `${readabilityAnalysis.score}%`,
-                     backgroundColor: `var(--progress-readability)`,
-                   }}
-                 />
-               </div>
-              <div className="flex justify-between items-baseline gap-2">
-                <div className="flex items-center gap-1 min-w-0">
-                  <InfoTooltip
-                    content={`80点以上：やさしい
-50-79点：普通
-50点未満：難しい`}
-                    className="text-sm text-foreground-secondary whitespace-nowrap"
-                  >
-                    難易度
-                  </InfoTooltip>
-                </div>
-                <span className="text-sm font-semibold text-foreground flex-shrink-0">
-                  {getReadabilityLevelLabel(readabilityAnalysis.level)}
-                </span>
-              </div>
-              <div className="pt-1 border-t border-border space-y-1">
-                <div className="flex justify-between items-baseline text-xs gap-2">
-                  <div className="flex items-center gap-1 min-w-0">
-                    <InfoTooltip
-                      content="1文あたりの平均文字数。40字以上は長め、20字以下は短めの文章"
-                      className="text-foreground-tertiary whitespace-nowrap"
-                    >
-                      一文平均
-                    </InfoTooltip>
-                  </div>
-                  <span className="text-foreground flex-shrink-0 text-sm">{readabilityAnalysis.avgSentenceLength}字/文</span>
-                </div>
-                <div className="flex justify-between items-baseline text-xs gap-2">
-                  <div className="flex items-center gap-1 min-w-0">
-                    <InfoTooltip
-                      content="句読点（、。）の間の平均文字数。15字以下が読みやすい"
-                      className="text-foreground-tertiary whitespace-nowrap"
-                    >
-                      句読点間隔
-                    </InfoTooltip>
-                  </div>
-                  <span className="text-foreground flex-shrink-0 text-sm">{readabilityAnalysis.avgPunctuationSpacing}字</span>
-                </div>
-              </div>
-            </div>
-          </div>
-        )}
-
-        {/* 2. 文字数内訳 */}
-        <div className="bg-background-secondary rounded-lg p-4 border border-border">
-          <h4 className="text-xs font-medium text-foreground-tertiary uppercase tracking-wide mb-3">
-             文字数
-          </h4>
-          <div className="space-y-1.5">
-            <div className="flex justify-between items-baseline gap-2">
-              <div className="flex items-center gap-1 min-w-0">
-                <InfoTooltip
-                  content="空白・改行を含むすべての文字数（原稿用紙換算の基準）"
-                  className="text-sm text-foreground-secondary whitespace-nowrap"
-                >
-                  総字数
-                </InfoTooltip>
-              </div>
-              <span className="text-base font-semibold text-foreground flex-shrink-0">{activeCharCount.toLocaleString()}</span>
-            </div>
-            <div className="flex justify-between items-baseline gap-2">
-              <div className="flex items-center gap-1 min-w-0">
-                <InfoTooltip
-                  content="文末の句点（。）で区切られる文の数"
-                  className="text-sm text-foreground-secondary whitespace-nowrap"
-                >
-                  文章数
-                </InfoTooltip>
-              </div>
-              <span className="text-sm font-medium text-foreground flex-shrink-0">{sentenceCount}文</span>
-            </div>
-            {sentenceCount > 0 && (
-              <div className="flex justify-between items-baseline gap-2">
-                <div className="flex items-center gap-1 min-w-0">
-                  <InfoTooltip
-                    content="1文あたりの平均文字数。短いほど読みやすい"
-                    className="text-sm text-foreground-secondary whitespace-nowrap"
-                  >
-                    一文平均
-                  </InfoTooltip>
-                </div>
-                <span className="text-sm font-medium text-foreground flex-shrink-0">
-                  {readabilityAnalysis ? `${readabilityAnalysis.avgSentenceLength}字/文` : '-'}
-                </span>
-              </div>
-            )}
-            <div className="flex justify-between items-baseline gap-2">
-              <div className="flex items-center gap-1 min-w-0">
-                <InfoTooltip
-                  content="句読点・記号を除いた本文のみの文字数"
-                  className="text-sm text-foreground-secondary whitespace-nowrap"
-                >
-                  本文字数
-                </InfoTooltip>
-              </div>
-              <span className="text-sm font-medium text-foreground flex-shrink-0">{pureTextCount.toLocaleString()}</span>
-            </div>
-            <div className="flex justify-between items-baseline gap-2">
-              <div className="min-w-0">
-                <div className="flex items-center gap-1">
-                  <InfoTooltip
-                    content="記号・句読点の割合。15%超：会話文中心、8%未満：地の文中心"
-                    className="text-sm text-foreground-secondary whitespace-nowrap"
-                  >
-                    約物比率
-                  </InfoTooltip>
-                </div>
-                <div className="text-xs text-foreground-tertiary">({styleHint})</div>
-              </div>
-              <span className="text-sm font-medium text-foreground flex-shrink-0">
-                {punctuationRatio}%
-              </span>
-            </div>
-          </div>
-        </div>
-
-        {/* 3. 文字種内訳 (Character Type Analysis) */}
-        {charTypeAnalysis && !isSelection && (
-          <div className="bg-background-secondary rounded-lg p-4 border border-border">
-            <h4 className="text-xs font-medium text-foreground-tertiary uppercase tracking-wide mb-3">
-               文字種別
-            </h4>
-            <div className="space-y-1.5">
-              <div className="flex justify-between items-baseline gap-2">
-                <div className="flex items-center gap-1 min-w-0">
-                  <InfoTooltip
-                    content="漢字の使用数と割合。一般的に20-30%が読みやすい"
-                    className="text-sm text-foreground-secondary whitespace-nowrap"
-                  >
-                    漢字
-                  </InfoTooltip>
-                </div>
-                <span className="text-sm font-medium text-foreground flex-shrink-0">
-                  {charTypeAnalysis.kanji} 字 {charUsageRates ? `(${charUsageRates.kanjiRate.toFixed(1)}%)` : ''}
-                </span>
-              </div>
-              <div className="flex justify-between items-baseline gap-2">
-                <div className="flex items-center gap-1 min-w-0">
-                  <InfoTooltip
-                    content="ひらがなの使用数と割合。通常50-70%程度"
-                    className="text-sm text-foreground-secondary whitespace-nowrap"
-                  >
-                    ひらがな
-                  </InfoTooltip>
-                </div>
-                <span className="text-sm font-medium text-foreground flex-shrink-0">
-                  {charTypeAnalysis.hiragana} 字 {charUsageRates ? `(${charUsageRates.hiraganaRate.toFixed(1)}%)` : ''}
-                </span>
-              </div>
-              <div className="flex justify-between items-baseline gap-2">
-                <div className="flex items-center gap-1 min-w-0">
-                  <InfoTooltip
-                    content="カタカナの使用数と割合。外来語や擬音語に使用"
-                    className="text-sm text-foreground-secondary whitespace-nowrap"
-                  >
-                    カタカナ
-                  </InfoTooltip>
-                </div>
-                <span className="text-sm font-medium text-foreground flex-shrink-0">
-                  {charTypeAnalysis.katakana} 字 {charUsageRates ? `(${charUsageRates.katakanaRate.toFixed(1)}%)` : ''}
-                </span>
-              </div>
-              {charTypeAnalysis.other > 0 && (
-                <div className="flex justify-between items-baseline">
-                  <span className="text-sm text-foreground-secondary">その他</span>
-                  <span className="text-sm font-medium text-foreground">{charTypeAnalysis.other}字</span>
-                </div>
-              )}
-               {/* 文字種の分布バー */}
-              <div className="pt-1 space-y-1">
-                <div className="h-6 flex rounded-md overflow-hidden bg-background border border-border-secondary">
-                  {charTypeAnalysis.total > 0 && (
-                    <>
-                      {charTypeAnalysis.kanji > 0 && (
-                        <div 
-                          className="flex items-center justify-center text-white text-xs font-semibold"
-                          style={{ 
-                            width: `${(charTypeAnalysis.kanji / charTypeAnalysis.total) * 100}%`,
-                            backgroundColor: `var(--progress-kanji)`
-                          }}
-                          title={`漢字: ${charTypeAnalysis.kanji}`}
-                        />
-                      )}
-                      {charTypeAnalysis.hiragana > 0 && (
-                        <div 
-                          className="flex items-center justify-center text-white text-xs font-semibold"
-                          style={{ 
-                            width: `${(charTypeAnalysis.hiragana / charTypeAnalysis.total) * 100}%`,
-                            backgroundColor: `var(--progress-hiragana)`
-                          }}
-                          title={`ひらがな: ${charTypeAnalysis.hiragana}`}
-                        />
-                      )}
-                      {charTypeAnalysis.katakana > 0 && (
-                        <div 
-                          className="flex items-center justify-center text-white text-xs font-semibold"
-                          style={{ 
-                            width: `${(charTypeAnalysis.katakana / charTypeAnalysis.total) * 100}%`,
-                            backgroundColor: `var(--progress-katakana)`
-                          }}
-                          title={`カタカナ: ${charTypeAnalysis.katakana}`}
-                        />
-                      )}
-                      {charTypeAnalysis.other > 0 && (
-                        <div 
-                          className="flex items-center justify-center text-white text-xs font-semibold"
-                          style={{ 
-                            width: `${(charTypeAnalysis.other / charTypeAnalysis.total) * 100}%`,
-                            backgroundColor: `var(--progress-other)`
-                          }}
-                          title={`その他: ${charTypeAnalysis.other}`}
-                        />
-                      )}
-                    </>
-                  )}
-                </div>
-                <div className="grid grid-cols-4 gap-2 text-xs">
-                  <div className="flex items-center gap-1">
-                    <div className="w-2 h-2 rounded-full" style={{ backgroundColor: `var(--progress-kanji)` }} />
-                    <span className="text-foreground-tertiary">漢字</span>
-                  </div>
-                  <div className="flex items-center gap-1">
-                    <div className="w-2 h-2 rounded-full" style={{ backgroundColor: `var(--progress-hiragana)` }} />
-                    <span className="text-foreground-tertiary">ひらがな</span>
-                  </div>
-                  <div className="flex items-center gap-1">
-                    <div className="w-2 h-2 rounded-full" style={{ backgroundColor: `var(--progress-katakana)` }} />
-                    <span className="text-foreground-tertiary">カタカナ</span>
-                  </div>
-                  <div className="flex items-center gap-1">
-                    <div className="w-2 h-2 rounded-full" style={{ backgroundColor: `var(--progress-other)` }} />
-                    <span className="text-foreground-tertiary">その他</span>
-                  </div>
-                </div>
-             </div>
-           </div>
-         </div>
-       )}
-
-        {/* 4. 段落構成 */}
-        <div className="bg-background-secondary rounded-lg p-4 border border-border">
-          <h4 className="text-xs font-medium text-foreground-tertiary uppercase tracking-wide mb-3">
-             段落
-          </h4>
-          <div className="space-y-1.5">
-            <div className="flex justify-between items-baseline gap-2">
-              <div className="flex items-center gap-1 min-w-0">
-                <InfoTooltip
-                  content="改行で区切られる段落の総数"
-                  className="text-sm text-foreground-secondary whitespace-nowrap"
-                >
-                  段落数
-                </InfoTooltip>
-              </div>
-              <span className="text-base font-semibold text-foreground flex-shrink-0">{paragraphCount}</span>
-            </div>
-            <div className="flex justify-between items-baseline gap-2">
-              <div className="flex items-center gap-1 min-w-0">
-                <InfoTooltip
-                  content="段落構成の傾向を見るための指標です。良し悪しを示すものではなく、文章設計を振り返るための参考値です。"
-                  className="text-sm text-foreground-secondary whitespace-nowrap"
-                >
-                  一段落平均
-                </InfoTooltip>
-              </div>
-              <span className="text-sm font-medium text-foreground flex-shrink-0">{avgParagraphLength}字/段</span>
-            </div>
-           {paragraphWarning && (
-             <div className="mt-2">
-               <div className="h-px bg-border" />
-                <small className="mt-2 block text-[10px] text-foreground/50">
-                  補足: {paragraphWarning}
-                </small>
-             </div>
-           )}
-         </div>
-       </div>
-
-        {/* 5. 読了時間（目安） */}
-        <div className="bg-background-secondary rounded-lg p-4 border border-border">
-          <h4 className="text-xs font-medium text-foreground-tertiary uppercase tracking-wide mb-3">
-             読了時間
-          </h4>
-          <div className="space-y-1.5">
-            <div className="flex justify-between items-baseline gap-2">
-              <div className="flex items-center gap-1 min-w-0">
-                <InfoTooltip
-                  content="分速900字で計算"
-                  className="text-sm text-foreground-secondary whitespace-nowrap"
-                >
-                  速読時
-                </InfoTooltip>
-              </div>
-              <span className="text-sm font-medium text-foreground flex-shrink-0">{fastReadTime}</span>
-            </div>
-            <div className="flex justify-between items-baseline gap-2">
-              <div className="flex items-center gap-1 min-w-0">
-                <InfoTooltip
-                  content="通常の読書速度（分速500字、日本語の平均的な速度）"
-                  className="text-sm text-foreground-secondary whitespace-nowrap"
-                >
-                  通常時
-                </InfoTooltip>
-              </div>
-              <span className="text-sm font-medium text-foreground flex-shrink-0">{normalReadTime}</span>
-            </div>
-            <div className="flex justify-between items-baseline gap-2">
-              <div className="flex items-center gap-1 min-w-0">
-                <InfoTooltip
-                  content="じっくり読む速度（分速250字で計算）"
-                  className="text-sm text-foreground-secondary whitespace-nowrap"
-                >
-                  精読時
-                </InfoTooltip>
-              </div>
-              <span className="text-sm font-medium text-foreground flex-shrink-0">{deepReadTime}</span>
-            </div>
-          </div>
-        </div>
-
-
-
-     </div>
-   );
-}
-
-
