+++ components/inspector/index.ts
@@ -1,0 +1,2 @@
+export { default } from "../Inspector";
+export type { InspectorProps, Tab, SeverityFilter, EnrichedLintIssue } from "./types";
